pkg/repo: Move repo validation into package
